+++ cli/src/lib.rs
@@ -521,16 +521,16 @@ fn init(cfg_override: &ConfigOverride, name: String, javascript: bool, no_git: b
     );
     cfg.programs.insert(Cluster::Localnet, localnet);
     let toml = cfg.to_string();
-    let mut file = File::create("Anchor.toml")?;
-    file.write_all(toml.as_bytes())?;
+    fs::write("Anchor.toml", toml)?;
 
     // Build virtual manifest.
-    let mut virt_manifest = File::create("Cargo.toml")?;
-    virt_manifest.write_all(template::virtual_manifest().as_bytes())?;
+    fs::write("Cargo.toml", template::virtual_manifest())?;
 
     // Initialize .gitignore file
-    let mut virt_manifest = File::create(".gitignore")?;
-    virt_manifest.write_all(template::git_ignore().as_bytes())?;
+    fs::write(".gitignore", template::git_ignore())?;
+
+    // Initialize .prettierignore file
+    fs::write(".prettierignore", template::prettier_ignore())?;
 
     // Build the program.
     fs::create_dir("programs")?;
@@ -1828,6 +1828,9 @@ fn test(
             )?;
         }
 
+        let root = cfg.path().parent().unwrap().to_owned();
+        cfg.add_test_config(root)?;
+
         // Run the deploy against the cluster in two cases:
         //
         // 1. The cluster is not localnet.
@@ -3050,14 +3053,17 @@ fn localnet(
 //
 // The closure passed into this function must never change the working directory
 // to be outside the workspace. Doing so will have undefined behavior.
-fn with_workspace<R>(cfg_override: &ConfigOverride, f: impl FnOnce(&WithPath<Config>) -> R) -> R {
+fn with_workspace<R>(
+    cfg_override: &ConfigOverride,
+    f: impl FnOnce(&mut WithPath<Config>) -> R,
+) -> R {
     set_workspace_dir_or_exit();
 
-    let cfg = Config::discover(cfg_override)
+    let mut cfg = Config::discover(cfg_override)
         .expect("Previously set the workspace dir")
         .expect("Anchor.toml must always exist");
 
-    let r = f(&cfg);
+    let r = f(&mut cfg);
 
     set_workspace_dir_or_exit();