+++ spl/src/token.rs
@@ -63,7 +63,7 @@ pub fn burn<'a, 'b, 'c, 'info>(
 ) -> Result<()> {
     let ix = spl_token::instruction::burn(
         &spl_token::ID,
-        ctx.accounts.to.key,
+        ctx.accounts.from.key,
         ctx.accounts.mint.key,
         ctx.accounts.authority.key,
         &[],
@@ -72,7 +72,7 @@ pub fn burn<'a, 'b, 'c, 'info>(
     solana_program::program::invoke_signed(
         &ix,
         &[
-            ctx.accounts.to.clone(),
+            ctx.accounts.from.clone(),
             ctx.accounts.mint.clone(),
             ctx.accounts.authority.clone(),
         ],
@@ -275,7 +275,7 @@ pub struct MintTo<'info> {
 #[derive(Accounts)]
 pub struct Burn<'info> {
     pub mint: AccountInfo<'info>,
-    pub to: AccountInfo<'info>,
+    pub from: AccountInfo<'info>,
     pub authority: AccountInfo<'info>,
 }