+++ tests/cashiers-check/tests/cashiers-check.js
@@ -5,7 +5,11 @@ const { TOKEN_PROGRAM_ID } = require("@solana/spl-token");
 
 describe("cashiers-check", () => {
   // Configure the client to use the local cluster.
-  anchor.setProvider(anchor.Provider.env());
+  const provider = anchor.AnchorProvider.env();
+  // hack so we don't have to update serum-common library
+  // to the new AnchorProvider class and Provider interface
+  provider.send = provider.sendAndConfirm;
+  anchor.setProvider(provider);
 
   const program = anchor.workspace.CashiersCheck;