+++ tests/cfo/tests/cfo.js
@@ -23,9 +23,13 @@ const SYSVAR_INSTRUCTIONS_PUBKEY = new PublicKey(
 const FEES = "6160355581";
 
 describe("cfo", () => {
-  anchor.setProvider(anchor.Provider.env());
+  const provider = anchor.AnchorProvider.env();
+  anchor.setProvider(provider);
 
   const program = anchor.workspace.Cfo;
+  // hack so we don't have to update serum-common library
+  // to the new AnchorProvider class and Provider interface
+  program.provider.send = provider.sendAndConfirm;
   const sweepAuthority = program.provider.wallet.publicKey;
   let officer, srmVault, usdcVault, bVault, stake, treasury;
   let officerBump, srmBump, usdcBump, bBump, stakeBump, treasuryBump;