+++ tests/cfo/tests/utils/stake.js
@@ -5,7 +5,10 @@ const utils = require("../../deps/stake/tests/utils");
 
 const lockup = anchor.workspace.Lockup;
 const registry = anchor.workspace.Registry;
-const provider = anchor.Provider.env();
+const provider = anchor.AnchorProvider.env();
+// hack so we don't have to update serum-common library
+// to the new AnchorProvider class and Provider interface
+provider.send = provider.sendAndConfirm;
 
 let lockupAddress = null;
 let mint = null;