+++ tests/errors/tests/errors.ts
@@ -64,7 +64,7 @@ const withLogTest = async (callback, expectedLogs) => {
 
 describe("errors", () => {
   // Configure the client to use the local cluster.
-  const localProvider = anchor.Provider.local();
+  const localProvider = anchor.AnchorProvider.local();
   localProvider.opts.skipPreflight = true;
   // processed failed tx do not result in AnchorErrors in the client
   // because we cannot get logs for them (only through overkill `onLogs`)
@@ -279,7 +279,7 @@ describe("errors", () => {
           data: program.coder.instruction.encode("signer_error", {}),
         })
       );
-      await program.provider.send(tx);
+      await program.provider.sendAndConfirm(tx);
       assert.ok(false);
     } catch (err) {
       anchor.getProvider().connection.removeOnLogsListener(listener);